Testes/silero-STT/requirements.txt
torch
torchaudio
soundfile
omegaconf
intel-tensorflow
tensorflow_hub